# docs/swagger.yaml
basePath: /
definitions:
  dto.AddItemRequest:
    properties:
      product_id:
        description: UserID    uint    `json:"user_id,omitempty"`
        type: string
      quantity:
        type: integer
      variant_id:
        type: string
    required:
    - product_id
    - quantity
    type: object
...
  dto.BulkUpdateRequest:
    properties:
      items:
        items:
          properties:
            product_id:
              type: string
            quantity:
              type: integer
            variant_id:
              type: string
          required:
          - product_id
          - quantity
          type: object
        type: array
      userID:
        type: integer
    type: object
...
  dto.CartItemResponse:
    properties:
      attributes:
        additionalProperties: {}
        type: object
      id:
        type: integer
      product:
        $ref: '#/definitions/dto.ProductResponse'
      product_id:
        type: string
      quantity:
        type: integer
      subtotal:
        type: number
      variant:
        allOf:
        - $ref: '#/definitions/dto.VariantResponse'
        description: Embed for display
      variant_id:
        type: string
    type: object
  dto.CartResponse:
    properties:
      created_at:
        description: Added
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/dto.CartItemResponse'
        type: array
      status:
        $ref: '#/definitions/models.CartStatus'
      total:
        type: number
      updated_at:
        description: Added
        type: string
      user_id:
        type: integer
    type: object
...
  dto.InventoryResponse:
    properties:
      backorder_allowed:
        type: boolean
      id:
        type: string
      low_stock_threshold:
        type: integer
      quantity:
        type: integer
      reserved_quantity:
        type: integer
    type: object
...
  dto.MediaResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      product_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
...
  dto.ProductResponse:
    properties:
      base_price:
        type: number
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      discount:
        minimum: 0
        type: number
      discount_type:
        enum:
        - fixed
        - percentage
        - ""
        type: string
      final_price:
        type: number
      id:
        type: string
      media:
        items:
          $ref: '#/definitions/dto.MediaResponse'
        type: array
      merchant_id:
        type: string
      name:
        type: string
      simple_inventory:
        allOf:
        - $ref: '#/definitions/dto.InventoryResponse'
        description: For simple products
      sku:
        type: string
      updated_at:
        type: string
      variants:
        items:
          $ref: '#/definitions/dto.VariantResponse'
        type: array
    type: object
...
  dto.UpdateItemRequest:
    properties:
      quantity:
        description: Omitempty for PATCH
        type: integer
    type: object
...
  dto.VariantResponse:
    properties:
      attributes:
        additionalProperties:
          type: string
        type: object
      created_at:
        type: string
      discount:
        minimum: 0
        type: number
      discount_type:
        enum:
        - fixed
        - percentage
        - ""
        type: string
      final_price:
        type: number
      id:
        type: string
      inventory:
        $ref: '#/definitions/dto.InventoryResponse'
      is_active:
        type: boolean
      price_adjustment:
        type: number
      product_id:
        type: string
      sku:
        type: string
      total_price:
        type: number
      updated_at:
        type: string
    type: object
...
  models.CartStatus:
    enum:
    - Active
    - Abandoned
    - Converted
    type: string
    x-enum-varnames:
    - CartStatusActive
    - CartStatusAbandoned
    - CartStatusConverted
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for customer and merchant authentication in a multivendor platform
  license:
    name: MIT
  termsOfService: http://example.com/terms/
  title: Multivendor API
  version: "1.0"
paths:
...
  /cart:
    get:
      description: Retrieves the user's active cart with items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CartResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get active cart
      tags:
      - Cart
  /cart/bulk:
    post:
      consumes:
      - application/json
      description: Adds multiple items to the user's active cart in one request
      parameters:
      - description: Bulk items details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.BulkUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CartResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Bulk add items to cart
      tags:
      - Cart
  /cart/clear:
    post:
      description: Clears all items from the user's active cart
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            properties:
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Clear the cart
      tags:
      - Cart
  /cart/items:
    post:
      consumes:
      - application/json
      description: Adds a product (with optional variant) to user's active cart
      parameters:
      - description: Item details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AddItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CartResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - Cart
  /cart/items/{id}:
    delete:
      description: Removes a cart item by ID
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CartResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove cart item
      tags:
      - Cart
    get:
      description: Retrieves a specific cart item
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CartItemResponse'
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get cart item by ID
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Updates the quantity of a cart item
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: string
      - description: New quantity
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CartResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update cart item quantity
      tags:
      - Cart
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"